+++ CHANGELOG.md
@@ -58,8 +58,10 @@
 ### ⬆️ Improved
 
 ### ✅ Added
+- `ChannelEventsHandler` is added to `QueryChannelsController` to handle updating channel list logic after receiving events. You can provide custom `ChannelEventsHandler` through `ChannelListViewModel` or using `QueryChannelsController` directly.
 
 ### ⚠️ Changed
+- `QueryChannelsController::newChannelEventFilter` and `QueryChannelsController#checkFilterOnChannelUpdatedEvent` are now deprecated. See the deprecation log for more details.
 
 ### ❌ Removed